requirements.txt
streamlit
boto3
transformers
datasets
load_dotenv
torch
bitsandbytes
accelerate